add hcsr04 example

+++ user/utils.c
@@ -3,7 +3,7 @@
  * @author  Alientek, Miaow
  * @version 0.2.0
  * @date    2019/07/22
- * @brief   
+ * @brief
  *          This file provides utilities:
  *              1. Delay functions
  *              2. Serialport on UART1. Functions from stdio.h are avaliable.
@@ -26,20 +26,20 @@
  *              2. Avoid to commit several features in one commit.
  *              3. Make your modification compact - don't reformat source code in your request.
  */
- 
+
 #include "utils.h"
 #include "system_stm32f4xx.h"
 
-/** @addtogroup UTILS
- * @{
- */
+ /** @addtogroup UTILS
+  * @{
+  */
 
 uint32_t AhbClock = 0; //!< AHB clock in Hz.
 uint32_t Apb1Clock = 0; //!< PCLK1(APB1 clock) in Hz.
 uint32_t Apb2Clock = 0; //!< PCLK2(APB2 clock) in Hz.
 
-static uint8_t fac_us = 0;
-static uint16_t fac_ms = 0;
+static float fac_us = 0;
+static float fac_ms = 0;
 #pragma import(__use_no_semihosting)             
 
 struct __FILE
@@ -65,7 +65,7 @@ void USART1_IRQHandler(void)
   {
     Res = USART_ReceiveData(USART1);//(USART1->DR);    //读取接收到的数据
 
-    if ((USART_RX_STA & 0x8000)==0)//接收未完成
+    if ((USART_RX_STA & 0x8000) == 0)//接收未完成
     {
       if (USART_RX_STA & 0x4000)//接收到了0x0d
       {
@@ -75,22 +75,22 @@ void USART1_IRQHandler(void)
           USART_RX_STA |= 0x8000;    //接收完成了 
       }
       else //还没收到0X0D
-      {    
+      {
         if (Res == 0x0d)
         {
           USART_RX_STA |= 0x4000;
         }
         else
         {
-          USART_RX_BUF[USART_RX_STA & 0X3FFF]=Res ;
+          USART_RX_BUF[USART_RX_STA & 0X3FFF] = Res;
           USART_RX_STA++;
           if (USART_RX_STA > (UTILS_RECEIEVE_LENTH - 1))
             USART_RX_STA = 0;//接收数据错误,重新开始接收      
         }
       }
-    }            
+    }
   }
-} 
+}
 #endif
 
 /**
@@ -104,84 +104,84 @@ void UTILS_UpdateClocks()
   uint32_t ahbClockTmp = (RCC->CFGR & ((uint32_t)15 << 4));
   uint32_t apb1ClockTmp = (RCC->CFGR & ((uint32_t)7 << 10));
   uint32_t apb2ClockTmp = (RCC->CFGR & ((uint32_t)7 << 13));
-  
+
   SystemCoreClockUpdate();
   switch (ahbClockTmp)
   {
-    case RCC_CFGR_HPRE_DIV1:
-      AhbClock = SystemCoreClock;
-      break;
-    case RCC_CFGR_HPRE_DIV2:
-      AhbClock = SystemCoreClock >> 1;
-      break;
-    case RCC_CFGR_HPRE_DIV4:
-      AhbClock = SystemCoreClock >> 2;
-      break;
-    case RCC_CFGR_HPRE_DIV8:
-      AhbClock = SystemCoreClock >> 3;
-      break;
-    case RCC_CFGR_HPRE_DIV16:
-      AhbClock = SystemCoreClock >> 4;
-      break;
-    case RCC_CFGR_HPRE_DIV64:
-      AhbClock = SystemCoreClock >> 6;
-      break;
-    case RCC_CFGR_HPRE_DIV128:
-      AhbClock = SystemCoreClock >> 7;
-      break;
-    case RCC_CFGR_HPRE_DIV256:
-      AhbClock = SystemCoreClock >> 8;
-      break;
-    case RCC_CFGR_HPRE_DIV512:
-      AhbClock = SystemCoreClock >> 9;
-      break;
-    default:
-      AhbClock = 0;
-      break;
+  case RCC_CFGR_HPRE_DIV1:
+    AhbClock = SystemCoreClock;
+    break;
+  case RCC_CFGR_HPRE_DIV2:
+    AhbClock = SystemCoreClock >> 1;
+    break;
+  case RCC_CFGR_HPRE_DIV4:
+    AhbClock = SystemCoreClock >> 2;
+    break;
+  case RCC_CFGR_HPRE_DIV8:
+    AhbClock = SystemCoreClock >> 3;
+    break;
+  case RCC_CFGR_HPRE_DIV16:
+    AhbClock = SystemCoreClock >> 4;
+    break;
+  case RCC_CFGR_HPRE_DIV64:
+    AhbClock = SystemCoreClock >> 6;
+    break;
+  case RCC_CFGR_HPRE_DIV128:
+    AhbClock = SystemCoreClock >> 7;
+    break;
+  case RCC_CFGR_HPRE_DIV256:
+    AhbClock = SystemCoreClock >> 8;
+    break;
+  case RCC_CFGR_HPRE_DIV512:
+    AhbClock = SystemCoreClock >> 9;
+    break;
+  default:
+    AhbClock = 0;
+    break;
   }
-  
+
   switch (apb1ClockTmp)
   {
-    case RCC_CFGR_PPRE1_DIV1:
-      Apb1Clock = AhbClock;
-      break;
-    case RCC_CFGR_PPRE1_DIV2:
-      Apb1Clock = AhbClock >> 1;
-      break;
-    case RCC_CFGR_PPRE1_DIV4:
-      Apb1Clock = AhbClock >> 2;
-      break;
-    case RCC_CFGR_PPRE1_DIV8:
-      Apb1Clock = AhbClock >> 3;
-      break;
-    case RCC_CFGR_PPRE1_DIV16:
-      Apb1Clock = AhbClock >> 4;
-      break;
-    default:
-      Apb1Clock = 0;
-      break;
+  case RCC_CFGR_PPRE1_DIV1:
+    Apb1Clock = AhbClock;
+    break;
+  case RCC_CFGR_PPRE1_DIV2:
+    Apb1Clock = AhbClock >> 1;
+    break;
+  case RCC_CFGR_PPRE1_DIV4:
+    Apb1Clock = AhbClock >> 2;
+    break;
+  case RCC_CFGR_PPRE1_DIV8:
+    Apb1Clock = AhbClock >> 3;
+    break;
+  case RCC_CFGR_PPRE1_DIV16:
+    Apb1Clock = AhbClock >> 4;
+    break;
+  default:
+    Apb1Clock = 0;
+    break;
   }
-  
+
   switch (apb2ClockTmp)
   {
-    case RCC_CFGR_PPRE2_DIV1:
-      Apb2Clock = AhbClock;
-      break;
-    case RCC_CFGR_PPRE2_DIV2:
-      Apb2Clock = AhbClock >> 1;
-      break;
-    case RCC_CFGR_PPRE2_DIV4:
-      Apb2Clock = AhbClock >> 2;
-      break;
-    case RCC_CFGR_PPRE2_DIV8:
-      Apb2Clock = AhbClock >> 3;
-      break;
-    case RCC_CFGR_PPRE2_DIV16:
-      Apb2Clock = AhbClock >> 4;
-      break;
-    default:
-      Apb2Clock = 0;
-      break;
+  case RCC_CFGR_PPRE2_DIV1:
+    Apb2Clock = AhbClock;
+    break;
+  case RCC_CFGR_PPRE2_DIV2:
+    Apb2Clock = AhbClock >> 1;
+    break;
+  case RCC_CFGR_PPRE2_DIV4:
+    Apb2Clock = AhbClock >> 2;
+    break;
+  case RCC_CFGR_PPRE2_DIV8:
+    Apb2Clock = AhbClock >> 3;
+    break;
+  case RCC_CFGR_PPRE2_DIV16:
+    Apb2Clock = AhbClock >> 4;
+    break;
+  default:
+    Apb2Clock = 0;
+    break;
   }
 }
 
@@ -191,86 +191,86 @@ void UTILS_UpdateClocks()
 void UTILS_InitDelay()
 {
   UTILS_UpdateClocks();
- 	SysTick_CLKSourceConfig(SysTick_CLKSource_HCLK_Div8);
-	fac_us = SystemCoreClock / 8000000;
-	fac_ms = (uint16_t)fac_us * 1000;
+  SysTick_CLKSourceConfig(SysTick_CLKSource_HCLK_Div8);
+  fac_us = (float)SystemCoreClock / 8000000.0f;
+  fac_ms = fac_us * 1000.0f;
 }
 
 /**
  * @brief Delay in microsecond.
- * @param time  Time in us. 
+ * @param time  Time in us.
  *              This parameter should be within the range of [0, 2 ^ 27 / SYSCLK(Hz)].
  * @note DO NOT use in OS.
  */
 void UTILS_DelayUs(uint32_t time)
 {
-	uint32_t temp;
-	SysTick->LOAD = time * fac_us;
-	SysTick->VAL = 0;
-	SysTick->CTRL |= SysTick_CTRL_ENABLE_Msk;
-	do
-	{
-		temp = SysTick->CTRL;
-	}while ((temp & 1) && !(temp & (1 << 16)));
-	SysTick->CTRL &= ~SysTick_CTRL_ENABLE_Msk;
-	SysTick->VAL = 0;
+  uint32_t temp;
+  SysTick->LOAD = (uint32_t)((float)time * fac_us);
+  SysTick->VAL = 0;
+  SysTick->CTRL |= SysTick_CTRL_ENABLE_Msk;
+  do
+  {
+    temp = SysTick->CTRL;
+  } while ((temp & 1) && !(temp & (1 << 16)));
+  SysTick->CTRL &= ~SysTick_CTRL_ENABLE_Msk;
+  SysTick->VAL = 0;
 }
 
 /**
  * @brief Delay in millisecond.
- * @param time  Time in ms. 
+ * @param time  Time in ms.
  *              This parameter should be within the range of [0, 134217720000 / SYSCLK(Hz)].
  * @note DO NOT use in OS.
  */
 static inline void UTILS_DelayXms(uint16_t nms)
-{	 		  	  
-	uint32_t temp;		   
-	SysTick->LOAD = (uint32_t)nms * fac_ms;
-	SysTick->VAL = 0;
-	SysTick->CTRL |= SysTick_CTRL_ENABLE_Msk;
-	do
-	{
-		temp=SysTick->CTRL;
-	}while ((temp & 1) && !(temp & (1 << 16)));
-	SysTick->CTRL &= ~SysTick_CTRL_ENABLE_Msk;
-	SysTick->VAL = 0;
-} 
+{
+  uint32_t temp;
+  SysTick->LOAD = (uint32_t)((float)nms * fac_ms);
+  SysTick->VAL = 0;
+  SysTick->CTRL |= SysTick_CTRL_ENABLE_Msk;
+  do
+  {
+    temp = SysTick->CTRL;
+  } while ((temp & 1) && !(temp & (1 << 16)));
+  SysTick->CTRL &= ~SysTick_CTRL_ENABLE_Msk;
+  SysTick->VAL = 0;
+}
 
 /**
  * @brief Delay in millisecond.
- * @param time  Time in ms. 
+ * @param time  Time in ms.
  * @note DO NOT use in OS.
  */
 void UTILS_DelayMs(uint16_t time)
-{	 	 
-	uint8_t repeat = time / 540;
-	uint16_t remain = time % 540;
-	while (repeat)
-	{
-		UTILS_DelayXms(540);
-		repeat--;
-	}
-	if (remain)
+{
+  uint8_t repeat = time / 540;
+  uint16_t remain = time % 540;
+  while (repeat)
+  {
+    UTILS_DelayXms(540);
+    repeat--;
+  }
+  if (remain)
     UTILS_DelayXms(remain);
 }
 
-void _sys_exit(int x) 
-{ 
-  x = x; 
+void _sys_exit(int x)
+{
+  x = x;
 }
 
-void _ttywrch(int ch) 
-{ 
-  ch = ch; 
+void _ttywrch(int ch)
+{
+  ch = ch;
 }
 
 /**
  * @brief Override fputc in stdlib.
  */
-int fputc(int ch, FILE *f)
-{     
-  while((USART1->SR & 0X40) == 0);
-  USART1->DR = (uint8_t)ch;      
+int fputc(int ch, FILE * f)
+{
+  while ((USART1->SR & 0X40) == 0);
+  USART1->DR = (uint8_t)ch;
   return ch;
 }
 
@@ -285,18 +285,18 @@ void UTILS_InitUart(uint32_t baudrate)
   NVIC_InitTypeDef NVIC_InitStructure;
 
   //clock
-  RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOA,ENABLE);
-  RCC_APB2PeriphClockCmd(RCC_APB2Periph_USART1,ENABLE);
+  RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOA, ENABLE);
+  RCC_APB2PeriphClockCmd(RCC_APB2Periph_USART1, ENABLE);
 
   //USART1 GPIO
-  GPIO_PinAFConfig(GPIOA,GPIO_PinSource9,GPIO_AF_USART1);
-  GPIO_PinAFConfig(GPIOA,GPIO_PinSource10,GPIO_AF_USART1);
+  GPIO_PinAFConfig(GPIOA, GPIO_PinSource9, GPIO_AF_USART1);
+  GPIO_PinAFConfig(GPIOA, GPIO_PinSource10, GPIO_AF_USART1);
   GPIO_InitStructure.GPIO_Pin = GPIO_Pin_9 | GPIO_Pin_10;
   GPIO_InitStructure.GPIO_Mode = GPIO_Mode_AF;
   GPIO_InitStructure.GPIO_Speed = GPIO_Speed_50MHz;
   GPIO_InitStructure.GPIO_OType = GPIO_OType_PP;
   GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_UP;
-  GPIO_Init(GPIOA,&GPIO_InitStructure);
+  GPIO_Init(GPIOA, &GPIO_InitStructure);
 
   //USART1
   USART_InitStructure.USART_BaudRate = baudrate;
@@ -309,20 +309,18 @@ void UTILS_InitUart(uint32_t baudrate)
 
   USART_Cmd(USART1, ENABLE);
 
-  //USART_ClearFlag(USART1, USART_FLAG_TC);
-
-  #if UTILS_USART_RX_ENABLE
+#if UTILS_USART_RX_ENABLE
   //USART1 receieve interrupt
-  USART_ITConfig(USART1, USART_IT_RXNE, ENABLE); 
+  USART_ITConfig(USART1, USART_IT_RXNE, ENABLE);
 
   //USART1 NVIC
   NVIC_InitStructure.NVIC_IRQChannel = USART1_IRQn;
-  NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority=3;
-  NVIC_InitStructure.NVIC_IRQChannelSubPriority =3;
+  NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 3;
+  NVIC_InitStructure.NVIC_IRQChannelSubPriority = 3;
   NVIC_InitStructure.NVIC_IRQChannelCmd = ENABLE;
   NVIC_Init(&NVIC_InitStructure);
-  #endif
+#endif
 }
-/** 
+/**
  * @}
  */
